requirements.txt
Django>=5.0,<6.0
django-cors-headers>=4.0
django-ckeditor>=6.0
Pillow>=10.0
python-dotenv>=1.0
psycopg2-binary>=2.9
